+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - `nstd_core_str[_mut]_from_raw_cstr[_with_null]` no longer panics if `cstr` is null.
 - `nstd_core_slice_mut_copy` now panics.
 ### `nstd.cstring`
+- `nstd_cstring_from_cstr[_unchecked]` now returns `NSTDOptionalCString`.
 - `nstd_cstring_clone` now returns `NSTDOptionalCString`.
 ### `nstd.heap_ptr`
 - `nstd_heap_ptr_[new[_zeroed]|clone]` now returns `NSTDOptionalHeapPtr`.
@@ -18,12 +19,14 @@
 ### `nstd.shared_ptr`
 - `nstd_shared_ptr_new[_zeroed]` now returns `NSTDOptionalSharedPtr`.
 ### `nstd.string`
+- `nstd_string_from_str` now returns `NSTDOptionalString`.
 - `nstd_string_clone` now returns `NSTDOptionalString`.
 ### `nstd.time`
 - `nstd.time` functions now take `NSTDTime` & `NSTDDuration` by value.
 ### `nstd.timed_mutex`
 - `nstd_timed_mutex_timed_lock` now takes `NSTDDuration` by value.
 ### `nstd.vec`
+- `nstd_vec_from_slice` now returns `NSTDOptionalVec`.
 - `nstd_vec_clone` now returns `NSTDOptionalVec`.
 
 # 0.6.0